refactor(qr-link): move verify_qr into decode module
verify_qr consumes the (version, hash) output of decode_qr_with_version
so they belong together. Use thiserror for UnsupportedVersion to reduce
boilerplate. Remove the separate verify.rs module.
